requirements.txt
dnspython==2.6.1
pythonping==1.1.4
requests==2.32.3
speedtest-cli==2.1.3
psutil
prometheus_client==0.20.0
PyYAML==6.0.1